HID functions to read back the status of an HID  controller (like MIDI SYSEX)

+++ src/controllers/controller.cpp
@@ -125,7 +125,8 @@ void Controller::receive(const QByteArray& data, mixxx::Duration timestamp) {
             } else {
                 spacer = QStringLiteral(" ");
             }
-            message += QString::number(data.at(i), 16)
+            // cast to quint8 to avoid that negative chars are for instance displayed as ffffffff instead of the desired ff
+            message += QString::number(static_cast<quint8>(data.at(i)), 16)
                                .toUpper()
                                .rightJustified(2, QChar('0')) +
                     spacer;
